refactoring nodeActivity (created class for converting details), martins snackbar fix

+++ app/src/main/java/freifunk/bremen/de/mobilemeshviewer/node/model/detail/TrafficBytes.java
@@ -26,7 +26,7 @@ public class TrafficBytes implements Parcelable {
     private long packets;
     @SerializedName("bytes")
     @Expose
-    private long bytes;
+    private double bytes;
 
     public TrafficBytes() {
     }
@@ -43,11 +43,11 @@ public class TrafficBytes implements Parcelable {
         this.packets = packets;
     }
 
-    public long getBytes() {
+    public double getBytes() {
         return bytes;
     }
 
-    public void setBytes(long bytes) {
+    public void setBytes(double bytes) {
         this.bytes = bytes;
     }
 
@@ -63,12 +63,12 @@ public class TrafficBytes implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeLong(bytes);
+        dest.writeDouble(bytes);
         dest.writeLong(packets);
     }
 
     private void readFromParcel(Parcel in) {
-        bytes = in.readLong();
+        bytes = in.readDouble();
         packets = in.readLong();
     }
 }
